build-logs: extract log tailing from `hydra-ws` into a crate
`hydra-ws` follows a build step's log file as the queue runner writes
it, knows where that file is from the step's derivation, and parses
the `step_finished` / `build_finished` notifications that tell it
when to stop. None of that is specific to websockets: any service
that shows a log live needs the same three things, and a second one
is about to. Move them to `subprojects/crates/build-logs` so the
services agree on all three, with the tailer's tests coming along.

`hydra-ws` keeps its own subscription bookkeeping and message types;
only the log-following core moves.
