+++ Cargo.toml
@@ -122,6 +122,7 @@ zstd = "0.13"
 
 # Local workspace crates
 binary-cache        = { path = "subprojects/crates/binary-cache" }
+build-logs          = { path = "subprojects/crates/build-logs" }
 daemon-client-utils = { path = "subprojects/crates/daemon-client-utils" }
 db                  = { path = "subprojects/crates/db" }
 hydra-proto         = { path = "subprojects/crates/proto" }
