+++ subprojects/hydra-ws/src/pg_listener.rs
@@ -5,6 +5,8 @@ use futures_util::StreamExt as _;
 
 use crate::state::State;
 
+use build_logs::notify::{parse_build_finished_payload, parse_step_finished_payload};
+
 const CHANNEL_STEP_FINISHED: &str = "step_finished";
 const CHANNEL_BUILD_FINISHED: &str = "build_finished";
 
@@ -67,60 +69,5 @@ pub async fn run_event_listener(state: Arc<State>) -> Result<(), ListenerError> 
                 tracing::warn!(channel, "unknown PG notification channel");
             }
         }
-    }
-}
-
-fn parse_step_finished_payload(payload: &str) -> Option<(u64, u64)> {
-    let parts: Vec<&str> = payload.split('\t').collect();
-    if parts.len() < 3 {
-        return None;
-    }
-    Some((parts[0].parse().ok()?, parts[1].parse().ok()?))
-}
-
-fn parse_build_finished_payload(payload: &str) -> Vec<u64> {
-    payload.split('\t').filter_map(|s| s.parse().ok()).collect()
-}
-
-#[cfg(test)]
-mod tests {
-    #![allow(clippy::unwrap_used, clippy::expect_used)]
-
-    use super::{parse_build_finished_payload, parse_step_finished_payload};
-
-    #[test]
-    fn parse_step_finished_valid() {
-        let result = parse_step_finished_payload("42\t5\tlogfile");
-        assert_eq!(result, Some((42, 5)));
-    }
-
-    #[test]
-    fn parse_step_finished_too_few_parts() {
-        let result = parse_step_finished_payload("42");
-        assert!(result.is_none());
-    }
-
-    #[test]
-    fn parse_step_finished_invalid_ids() {
-        let result = parse_step_finished_payload("abc\t5\tlog");
-        assert!(result.is_none());
-    }
-
-    #[test]
-    fn parse_build_finished_valid() {
-        let result = parse_build_finished_payload("42");
-        assert_eq!(result, vec![42]);
-    }
-
-    #[test]
-    fn parse_build_finished_includes_dependents() {
-        let result = parse_build_finished_payload("42\t43\t44");
-        assert_eq!(result, vec![42, 43, 44]);
-    }
-
-    #[test]
-    fn parse_build_finished_invalid_payload_is_empty() {
-        assert!(parse_build_finished_payload("").is_empty());
-        assert!(parse_build_finished_payload("not-an-id").is_empty());
     }
 }
